chore: standard renderers moved, dependencies added, documented
