+++ packages/furo-ui5/README.md
@@ -31,6 +31,7 @@ For seamless integration into the Furo environment, we have slightly extended so
 - [furo-ui5-data-display](https://components.furo.pro/?t=FuroUi5DataDisplay) a display field component 
 - [furo-ui5-button](https://components.furo.pro/?t=FuroUi5Button) a simple button with convenience functions 
 - [furo-ui5-busyindicator](https://components.furo.pro/?t=FuroUi5Busyindicator) component to signal that some operation is going on
+- [furo-ui5-data-table](https://components.furo.pro/?t=FuroUi5DataTable) component to display data in tabular form
 
 ### Compositions
 - [furo-ui5-pagination](https://components.furo.pro/?t=FuroUI5Pagination) a pagination bar
