+++ packages/furo-ui5/src/furo-ui5-button-bar.js
@@ -51,7 +51,8 @@ class FuroUi5ButtonBar extends FuroButtonBar {
         }
 
         ::slotted(*) {
-          margin: var(--spacing-xs, 8px) var(--spacing-s, 16px)  var(--spacing-xs, 8px) var(--spacing-s, 16px);
+          margin: var(--spacing-xs, 8px) var(--spacing-s, 16px) var(--spacing-xs, 8px)
+            var(--spacing-s, 16px);
         }
       `
     );
