+++ packages/furo-ui5/src/furo-ui5-data-display.js
@@ -140,7 +140,7 @@ class FuroUi5DataDisplay extends FBP(LitElement) {
     this.binder.labelMappings = {
       error: '_error',
       disabled: 'disabled',
-      readonly: 'disabled'
+      readonly: 'disabled',
     };
 
     // set attributes to constrains mapping for furo.fat types
