+++ packages/furo-ui5/src/standard-type-renderers/display-furo-fat-string.js
@@ -1,19 +1,18 @@
 import { LitElement, html, css } from 'lit-element';
 
 /**
- * `display-string`
- * The display-string component displays a FieldNode of type `string` in read only mode.
+ * `display-furo-fat-string`
+ * The display-furo-fat-string component displays a FieldNode of type `furo.fat.String` in read only mode.
  *
  * Every display-xxx component should implement the following API:
  * - function: bindData(fieldNode){...}
  * - attribute: tabular-form (this attribute is set, if the component is inside of a furo-data-table. This attribute is only needed, if the styling inside of a table is different)
  *
- *
  * @summary
  * @customElement
- * @demo demo display-string Basic Usage
+ * @demo demo display-furo-fat-string Basic Usage
  */
-class DisplayString extends LitElement {
+class DisplayFuroFatString extends LitElement {
   constructor() {
     super();
     this._field = undefined;
@@ -35,17 +34,14 @@ class DisplayString extends LitElement {
         :host([value-state='Success']) {
           color: var(--sapPositiveColor, #107e3e);
         }
-
         :host([value-state='Informative']),
         :host([value-state='Information']) {
           color: var(--sapInformativeColor, #0a6ed1);
         }
-
         :host([value-state='Negative']),
         :host([value-state='Error']) {
           color: var(--sapNegativeColor, #b00);
         }
-
         :host([value-state='Critical']),
         :host([value-state='Warning']) {
           color: var(--sapCrticalColor, #e9730c);
@@ -60,6 +56,16 @@ class DisplayString extends LitElement {
    */
   bindData(fieldNode) {
     this._field = fieldNode;
+
+    /**
+     * Sets the attributes from the field node
+     */
+    Object.keys(this._field.attributes).forEach(key => {
+      if (!key.startsWith('_')) {
+        this.setAttribute(this._field.attributes[key]._name, this._field.attributes[key]._value);
+      }
+    });
+
     if (this._field) {
       this._field.addEventListener('field-value-changed', () => {
         this.requestUpdate();
@@ -73,7 +79,9 @@ class DisplayString extends LitElement {
    * @private
    */
   _getTemplate() {
-    return html` <span>${this._field._value}</span> `;
+    return html`
+      <span>${this._field._value.value}</span>
+    `;
   }
 
   /**
@@ -83,8 +91,10 @@ class DisplayString extends LitElement {
    */
   render() {
     // language=HTML
-    return html` ${this._getTemplate()} `;
+    return html`
+      ${this._getTemplate()}
+    `;
   }
 }
 
-window.customElements.define('display-string', DisplayString);
+window.customElements.define('display-furo-fat-string', DisplayFuroFatString);
