+++ packages/furo-ui5/src/standard-type-renderers/display-google-type-date.js
@@ -1,8 +1,12 @@
 import { LitElement, html, css } from 'lit-element';
 import { Env } from '@furo/framework/src/furo.js';
+
 /**
- * `display-double`
- * The display-double component displays a FieldNode of type `double` in read only mode.
+ * `display-google-type-date`
+ * The display-google-type-date component displays a FieldNode of type `google.type.Date` in read only mode.
+ *
+ * The component uses locale from the environment to display the date value accordingly.
+ * https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat/format
  *
  * Every display-xxx component should implement the following API:
  * - function: bindData(fieldNode){...}
@@ -10,12 +14,13 @@ import { Env } from '@furo/framework/src/furo.js';
  *
  * @summary
  * @customElement
- * @demo demo display-double Basic Usage
+ * @demo demo display-google-type-date Basic Usage
  */
-class DisplayDouble extends LitElement {
+class DisplayGoogleTypeDate extends LitElement {
   constructor() {
     super();
     this._field = undefined;
+    this._formattedDateString = '';
   }
 
   static get styles() {
@@ -24,11 +29,6 @@ class DisplayDouble extends LitElement {
       css`
         :host {
           display: block;
-          word-break: keep-all;
-        }
-        
-        :host([tabular-form]) {
-          text-align: right;
         }
 
         :host([hidden]) {
@@ -61,11 +61,34 @@ class DisplayDouble extends LitElement {
    */
   bindData(fieldNode) {
     this._field = fieldNode;
+
     if (this._field) {
       this._field.addEventListener('field-value-changed', () => {
+        this._formattedDateString = DisplayGoogleTypeDate._convertDateToString(this._field);
         this.requestUpdate();
       });
     }
+
+    this._formattedDateString = DisplayGoogleTypeDate._convertDateToString(this._field);
+  }
+
+  /**
+   * convert date object to String according to Intl DateTimeFormat
+   * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat/DateTimeFormat
+   * Example output: locale de-CH 31.12.2021
+   * @param date
+   * @returns {string}
+   * @private
+   */
+  static _convertDateToString(fieldNode) {
+    const date = new Date(
+      Date.UTC(fieldNode.year._value, fieldNode.month._value - 1, fieldNode.day._value, 0, 0, 0, 0),
+    );
+    return new Intl.DateTimeFormat([Env.locale, 'de-CH'], {
+      year: 'numeric',
+      month: '2-digit',
+      day: '2-digit',
+    }).format(date);
   }
 
   /**
@@ -74,8 +97,9 @@ class DisplayDouble extends LitElement {
    * @private
    */
   _getTemplate() {
-    this.displayValue = new Intl.NumberFormat(Env.locale, {}).format(this._field);
-    return html` <span>${this.displayValue}</span> `;
+    return html`
+      <span>${this._formattedDateString}</span>
+    `;
   }
 
   /**
@@ -85,8 +109,10 @@ class DisplayDouble extends LitElement {
    */
   render() {
     // language=HTML
-    return html` ${this._getTemplate()} `;
+    return html`
+      ${this._getTemplate()}
+    `;
   }
 }
 
-window.customElements.define('display-double', DisplayDouble);
+window.customElements.define('display-google-type-date', DisplayGoogleTypeDate);
