+++ packages/furo-ui5/src/standard-type-renderers/display-int32.js
@@ -2,8 +2,8 @@ import { LitElement, html, css } from 'lit-element';
 import { Env } from '@furo/framework/src/furo.js';
 
 /**
- * `display-int64`
- * The display-int64 component displays a FieldNode of type `int64` in read only mode.
+ * `display-int32`
+ * The display-int32 component displays a FieldNode of type `int32` in read only mode.
  *
  * The component uses locale from the environment to display the date value accordingly.
  * https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat
@@ -14,9 +14,9 @@ import { Env } from '@furo/framework/src/furo.js';
  *
  * @summary
  * @customElement
- * @demo demo display-int64 Basic Usage
+ * @demo demo display-int32 Basic Usage
  */
-class DisplayInt64 extends LitElement {
+class DisplayInt32 extends LitElement {
   constructor() {
     super();
     this._field = undefined;
@@ -79,7 +79,9 @@ class DisplayInt64 extends LitElement {
    */
   _getTemplate() {
     this.displayValue = new Intl.NumberFormat(Env.locale, {}).format(this._field);
-    return html` <span>${this.displayValue}</span> `;
+    return html`
+      <span>${this.displayValue}</span>
+    `;
   }
 
   /**
@@ -89,8 +91,10 @@ class DisplayInt64 extends LitElement {
    */
   render() {
     // language=HTML
-    return html` ${this._getTemplate()} `;
+    return html`
+      ${this._getTemplate()}
+    `;
   }
 }
 
-window.customElements.define('display-int64', DisplayInt64);
+window.customElements.define('display-int32', DisplayInt32);
