+++ packages/furo-ui5/src/standard-type-renderers/display-int64.js
@@ -1,8 +1,9 @@
 import { LitElement, html, css } from 'lit-element';
 import { Env } from '@furo/framework/src/furo.js';
+
 /**
- * `display-uint32`
- * The display-uint32 component displays a FieldNode of type `uint64` in read only mode.
+ * `display-int64`
+ * The display-int64 component displays a FieldNode of type `int64` in read only mode.
  *
  * The component uses locale from the environment to display the date value accordingly.
  * https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat
@@ -13,9 +14,9 @@ import { Env } from '@furo/framework/src/furo.js';
  *
  * @summary
  * @customElement
- * @demo demo display-uint32 Basic Usage
+ * @demo demo display-int64 Basic Usage
  */
-class DisplayUint32 extends LitElement {
+class DisplayInt64 extends LitElement {
   constructor() {
     super();
     this._field = undefined;
@@ -78,7 +79,9 @@ class DisplayUint32 extends LitElement {
    */
   _getTemplate() {
     this.displayValue = new Intl.NumberFormat(Env.locale, {}).format(this._field);
-    return html` <span>${this.displayValue}</span> `;
+    return html`
+      <span>${this.displayValue}</span>
+    `;
   }
 
   /**
@@ -88,8 +91,10 @@ class DisplayUint32 extends LitElement {
    */
   render() {
     // language=HTML
-    return html` ${this._getTemplate()} `;
+    return html`
+      ${this._getTemplate()}
+    `;
   }
 }
 
-window.customElements.define('display-uint32', DisplayUint32);
+window.customElements.define('display-int64', DisplayInt64);
