+++ packages/furo-ui5/src/standard-type-renderers/display-uint64.js
@@ -1,9 +1,11 @@
 import { LitElement, html, css } from 'lit-element';
 import { Env } from '@furo/framework/src/furo.js';
-
 /**
- * `display-float`
- * The display-float component displays a FieldNode of type `float` in read only mode.
+ * `display-uint64`
+ * The display-uint64 component displays a FieldNode of type `uint64` in read only mode.
+ *
+ * The component uses locale from the environment to display the date value accordingly.
+ * https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Intl/NumberFormat
  *
  * Every display-xxx component should implement the following API:
  * - function: bindData(fieldNode){...}
@@ -11,9 +13,9 @@ import { Env } from '@furo/framework/src/furo.js';
  *
  * @summary
  * @customElement
- * @demo demo display-float Basic Usage
+ * @demo demo display-uint64 Basic Usage
  */
-class DisplayFloat extends LitElement {
+class DisplayUint64 extends LitElement {
   constructor() {
     super();
     this._field = undefined;
@@ -30,10 +32,6 @@ class DisplayFloat extends LitElement {
 
         :host([tabular-form]) {
           text-align: right;
-        }
-
-        :host([hidden]) {
-          display: none;
         }
 
         :host([value-state='Positive']),
@@ -76,7 +74,9 @@ class DisplayFloat extends LitElement {
    */
   _getTemplate() {
     this.displayValue = new Intl.NumberFormat(Env.locale, {}).format(this._field);
-    return html` <span>${this.displayValue}</span> `;
+    return html`
+      <span>${this.displayValue}</span>
+    `;
   }
 
   /**
@@ -86,8 +86,10 @@ class DisplayFloat extends LitElement {
    */
   render() {
     // language=HTML
-    return html` ${this._getTemplate()} `;
+    return html`
+      ${this._getTemplate()}
+    `;
   }
 }
 
-window.customElements.define('display-float', DisplayFloat);
+window.customElements.define('display-uint64', DisplayUint64);
